Fix teku beacon service start issue
When Teku migrated to standalone validator mode, it noticed that beacon service is not being started. Introduced explicit start task so that it make sure service is started. Moved restart into another task so restart only occurs if there has been any changes to the config file or service file